Use broccoli-caching-writer
We always see broccoli-styledown in the "slow-trees listing" for every
rebuild regardless of what files were changed.

This change updates to inherit from broccoli-caching-writer, which
simply symlinks (using symlink-or-copy package) from the last runs
cache directory if the input trees have not changed.

The result of this is that when changing files in a tree other than
those provided to broccoli-styledown the broccoli-styledown tree
rebuilds are nearly instant.